Implementation of SBU rejection of applications--

+++ Backend_UMR_Work_Program/DataModels/WP_OIL_CONDENSATE_PRODUCTION_ACTIVITIES_monthly_Activities_Pivotted_by_company_productionmonth_year_breakdown.cs
@@ -7,17 +7,17 @@ namespace Backend_UMR_Work_Program.Models
     {
         public string? CompanyName { get; set; }
         public string? Year_of_WP { get; set; }
-        public decimal? January { get; set; }
-        public decimal? February { get; set; }
-        public decimal? March { get; set; }
-        public decimal? April { get; set; }
-        public decimal? May { get; set; }
-        public decimal? June { get; set; }
-        public decimal? July { get; set; }
-        public decimal? August { get; set; }
-        public decimal? September { get; set; }
-        public decimal? October { get; set; }
-        public decimal? November { get; set; }
-        public decimal? December { get; set; }
+        public int? January { get; set; }
+        public int? February { get; set; }
+        public int? March { get; set; }
+        public int? April { get; set; }
+        public int? May { get; set; }
+        public int? June { get; set; }
+        public int? July { get; set; }
+        public int? August { get; set; }
+        public int? September { get; set; }
+        public int? October { get; set; }
+        public int? November { get; set; }
+        public int? December { get; set; }
     }
 }
